Add TRIM support for disk vdevs.
This adds TRIM (a.k.a UNMAP, DISCARD, hole punching) support for disk
vdevs. The original patch is from Pawel Jakub Dawidek
<pjd@FreeBSD.org> who wrote it for FreeBSD. Etienne Dechamps
<etienne.dechamps@ovh.net> ported it to ZFS On Linux.

The code builds a map of regions that were freed. On every write the
code consults the map and eventually removes ranges that were freed
before, but are now overwritten.

Freed blocks are not TRIMed immediately. There is a tunable that defines
how many txg we should wait with TRIMming freed blocks (64 by default).

There is a low priority thread that TRIMs ranges when the time comes.
During TRIM we keep in-flight ranges on a list to detect colliding
writes - we have to delay writes that collide with in-flight TRIMs in
case something will be reordered and write will reached the disk before
the TRIM. We don't have to do the same for in-flight writes, as
colliding writes just remove ranges to TRIM.

Most of the code stayed unchanged during the porting to Linux. The only
big change is in the vdev disk module, since the FreeBSD and Linux
interfaces for issuing discards to block devices is obviously different.
On FreeBSD it seems that issuing a DELETE request of any size is
sufficient; on Linux we have to be careful not to exceed maximum discard
limits. That's why we introduce a new vdev_disk_io_trim() function
inspired from the Linux blkdev_issue_discard() function and the
pre-existing vdev_disk_physio() function. The new function takes care of
splitting discard requests into smaller ones if necessary.

In theory, the code should work for main pool disk vdevs, slog disk
vdevs, L2ARC disk vdevs, and supports mirror and raidz. File vdevs are
not supported yet.

Note that the new feature is disabled by default (zfs_notrim=1). To use
it, you have to explictly set the module parameter "zfs_notrim" to "0".
Be aware that this code is largely untested and brings huge risks of
potential data corruption. Use at your own risk and expect data loss.