Fix BIO memory issues when using TRIM on disk vdevs.
The current code suffers from the following issues:

- When using DISCARD, the BIO is allocated with a number of vectors
  which have no purpose. It should be allocated with zero vectors.

- When using zfs_trim_zero and TRIMming large amounts of data, a
  large number of small page-sized (4K) BIOs is allocated. Because of
  this, some BIO allocations fail with ENOMEM.

This patch fixes these problems. The second issue is fixed by allowing
up to 128 zero vectors per BIO to reduce the number of BIOs being
allocated.

In addition, the code now makes sure that the DIO never needs to be
resized in the DISCARD case by calculating the needed size in advance.