Fix an issue with a NULL zio buf free on raidz.
When a FREE ZIO is issued on raidz, vdev_raidz_map_alloc() skips
allocating data for the raidz columns because it is not needed.

vdev_raidz_map_free(), however, has not been updated accordingly, and
still tries to free the column data. As a result, zio_buf_free() is
used with a NULL buffer, which corrupts the zio_buf_cache.

This patch fixes the issue by making vdev_raidz_map_free() do a NULL
check before freeing the column data.

This fixes the following panic when running zconfig.sh -c:

    SPLError: 17651:0:(spl-kmem.c:1936:spl_kmem_cache_alloc()) ASSERTION(obj) failed
    SPLError: 17651:0:(spl-kmem.c:1936:spl_kmem_cache_alloc()) SPL PANIC