Lock SCL_STATE while waiting for TRIM I/Os to complete.
The current code releases SCL_STATE while it is waiting for TRIM ZIOs
to complete. That's dangerous because the vdev configuration could
change while the I/O is in progress.

This patch makes sure SCL_STATE is held during the whole duration of
the I/O. This is consistent with the way e.g. zil_flush_vdevs() does
things.