Align trim map frees using vdev ashift.
Currently, frees are registered in the trim map using the physical size
of the block being freed. It is more efficient to use the allocated
size to increase the size of TRIM requests, and more importantly, to
make sure batching actually works for adjacent blocks.

This should be especially useful for high ashift pools.