Fix an integer underflow issue when freezing the SPA early.
Currently, we use the minimum of the last synced TXG and the freeze
TXG, minus trim_txg_limit, to decide on the last TXG to trim.

This substraction can underflow if the last synced TXG or the freeze
TXG is less than trim_txg_limit. We are supposed to check for that in
trim_map_commit(). Unfortunately, we only check for the last synced
TXG, and forgot to check for the freeze TXG. This means that if the
SPA is frozen early (before trim_txg_limit has passed), we will have
an underflow error.

With ztest, this results in pool corruption and the following error
message almost immediately after the test begins:

    ztest: ../../cmd/ztest/ztest.c:5733: Assertion `0 == spa_open(ztest_opts.zo_pool, &spa, ((char *)__func__)) (0x0 == 0x5)' failed.

This patch fixes the issue by checking for the freeze TXG in
trim_map_commit().