Issue #566: Add "-o ashift" to zpool add and zpool attach
When adding devices to an existing pool "ashift" property is auto-detected.
However, if this property was overridden at the pool creation time
(i.e. zpool create -o ashift=12 tank ...) this may not be what the user
wants.  This commit lets the user specify the value of "ashift" property
to be used with newly added drives. For example,

    zpool add -o ashift=12 tank disk1
    zpool attach -o ashift=12 tank disk1 disk2

Signed-off-by: Cyril Plisko <cyril.plisko@mountall.com>